added workflow for testing bactopia on github actions, tweaked gitlab ci

+++ bin/gh-actions/setup-bactopia-env.sh
@@ -6,9 +6,13 @@ set -x
 CONDA_PATH=${1:-"/opt/conda"}
 WORK_DIR=${2:-"/bactopia"}
 IS_GITHUB=${3:-"0"}
-CONDA_CMD="create -n bactopia"
+IS_GITLAB=${4:-"0"}
+ENV=${5:-"bactopia"}
+CONDA_CMD="create -n ${ENV}"
 if [[ "${IS_GITHUB}" == "1" ]]; then
   CONDA_CMD="install"
+elif [[ "${IS_GITLAB}" != "0" ]]; then
+  CONDA_CMD="create --prefix ${IS_GITLAB}"
 fi
 
 # Create environment
@@ -34,7 +38,7 @@ conda ${CONDA_CMD} --quiet -y -c conda-forge -c bioconda \
   wget
 
 # Setup variables
-BACTOPIA=${CONDA_PATH}/envs/bactopia
+BACTOPIA=${CONDA_PATH}/envs/${ENV}
 chmod 755 ${WORK_DIR}/bactopia ${WORK_DIR}/bin/helpers/*
 cp ${WORK_DIR}/bactopia ${WORK_DIR}/bin/helpers/* ${BACTOPIA}/bin
 VERSION=`${BACTOPIA}/bin/bactopia version | cut -d " " -f 2`
@@ -55,7 +59,7 @@ cp -r \
   ${BACTOPIA_SHARE}
 
 # Clean up
-if [[ "${IS_GITHUB}" == "0" ]]; then
+if [[ "${IS_GITHUB}" == "0" && "${IS_GITLAB}" == "0"]]; then
   rm -rf /bactopia
   conda clean -y -a
 fi
